requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3

torch>=2.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

openai>=1.12.0
python-dotenv>=1.0.0
